Changed OS processing to (script_windows.go, script_darwin.go) file.
