+++ script/script.go
@@ -3,9 +3,7 @@ package script
 import (
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"runtime"
 )
 
 // script
@@ -13,54 +11,21 @@ type Script interface {
 	Exec(command string) ([]byte, error)
 }
 
-type ScriptFactory func(data []byte) (Script, error)
 
-var files = map[string]string{
-	"windows": "files/win/iTunes.js",
-	"darwin":  "files/mac/ITunesTransport.scpt",
-}
+func createScriptFile(in,out string) (string, error) {
 
-var factory = map[string]ScriptFactory{
-	"windows": newWindowsScript,
-	"darwin":  newAppleScript,
-}
-
-func NewScript() (Script, error) {
-	data, err := Asset(files[runtime.GOOS])
+	data, err := Asset(in)
 	if err != nil {
-		return nil, err
-	}
-	return factory[runtime.GOOS](data)
-}
-
-func newAppleScript(data []byte) (Script, error) {
-	path, err := createScriptFile("mac_itunes.scpt", data)
-	if err != nil {
-		return nil, err
-	}
-	return AppleScript{
-		path: path,
-	}, nil
-}
-
-func newWindowsScript(data []byte) (Script, error) {
-	path, err := createScriptFile("win_itunes.js", data)
-	if err != nil {
-		return nil, err
+		return "", err
 	}
 
-	return WindowsScript{
-		path: path,
-	}, nil
-}
-
-func createScriptFile(name string, data []byte) (string, error) {
 	dir := os.TempDir()
-	path := filepath.Join(dir, name)
+	path := filepath.Join(dir, out)
 	if exists(path) {
 		return path, nil
 	}
-	err := ioutil.WriteFile(path, data, 744)
+
+	err = ioutil.WriteFile(path, data, 744)
 	if err != nil {
 		return "", err
 	}
@@ -72,18 +37,4 @@ func exists(filename string) bool {
 	return err == nil
 }
 
-type AppleScript struct {
-	path string
-}
 
-func (a AppleScript) Exec(command string) ([]byte, error) {
-	return exec.Command("osascript", a.path, command).Output()
-}
-
-type WindowsScript struct {
-	path string
-}
-
-func (a WindowsScript) Exec(command string) ([]byte, error) {
-	return exec.Command("cscript -e", a.path, command).Output()
-}
